chore: Move validation time stats to validation step
