+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/ValidationErrorReporter.java
@@ -28,6 +28,7 @@
 package org.hisp.dhis.tracker.validation;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.EnumMap;
 import java.util.HashSet;
 import java.util.List;
@@ -42,6 +43,7 @@ import lombok.Value;
 import org.hisp.dhis.tracker.TrackerIdSchemeParams;
 import org.hisp.dhis.tracker.TrackerType;
 import org.hisp.dhis.tracker.domain.TrackerDto;
+import org.hisp.dhis.tracker.report.Timing;
 import org.hisp.dhis.tracker.report.TrackerErrorCode;
 import org.hisp.dhis.tracker.report.TrackerErrorReport;
 import org.hisp.dhis.tracker.report.TrackerWarningReport;
@@ -68,6 +70,8 @@ public class ValidationErrorReporter
 
     TrackerIdSchemeParams idSchemes;
 
+    List<Timing> timings;
+
     @Getter( AccessLevel.PACKAGE )
     /*
      * Keeps track of all the invalid Tracker objects (i.e. objects with at
@@ -92,6 +96,7 @@ public class ValidationErrorReporter
         this.invalidDTOs = new EnumMap<>( TrackerType.class );
         this.idSchemes = idSchemes;
         this.isFailFast = failFast;
+        this.timings = new ArrayList<>();
     }
 
     /**
@@ -195,5 +200,27 @@ public class ValidationErrorReporter
     public boolean isInvalid( TrackerType trackerType, String uid )
     {
         return this.invalidDTOs.getOrDefault( trackerType, new HashSet<>() ).contains( uid );
+    }
+
+    public ValidationErrorReporter addTiming( Timing timing )
+    {
+        timings.add( timing );
+        return this;
+    }
+
+    public ValidationErrorReporter addTimings( List<Timing> timings )
+    {
+        this.timings.addAll( timings );
+        return this;
+    }
+
+    public List<Timing> getTimings()
+    {
+        return Collections.unmodifiableList( timings );
+    }
+
+    public boolean hasTimings()
+    {
+        return !timings.isEmpty();
     }
 }
